backend/pyproject.toml
[project]
name = "fastapi001"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aerich>=0.9.0",
    "aiomysql>=0.2.0",
    "fastapi>=0.115.12",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.11.4",
    "pypinyin>=0.54.0",
    "python-dotenv>=1.1.0",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "scipy>=1.13.1",
    "soundfile>=0.13.1",
    "tomlkit>=0.13.3",
    "tortoise-orm>=0.25.0",
    "uvicorn>=0.34.2",
]

[tool.aerich]
tortoise_orm = "core.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
